+++ packages/plugins/Gfi/src/components/Gfi.vue
@@ -16,17 +16,16 @@ import compare from 'just-compare'
 import { t } from 'i18next'
 import Vue from 'vue'
 import { mapActions, mapGetters, mapMutations } from 'vuex'
-import { GeoJsonProperties } from 'geojson'
 import { MoveHandleProperties } from '@polar/lib-custom-types'
 import Feature from './Feature.vue'
 import List from './List.vue'
 
 export default Vue.extend({
   name: 'GfiPlugin',
-  data: () => ({ clientWidth: 0 }),
   computed: {
     ...mapGetters(['moveHandle']),
     ...mapGetters('plugin/gfi', [
+      'currentProperties',
       'exportPropertyLayerKeys',
       'gfiContentComponent',
       'gfiConfiguration',
@@ -41,37 +40,7 @@ export default Vue.extend({
       return this.showList ? List : this.gfiContentComponent || Feature
     },
     contentProps(): object {
-      return this.showList
-        ? {}
-        : {
-            currentProperties: this.currentProperties,
-            clientWidth: this.clientWidth,
-            exportProperty: this.exportProperty,
-            showSwitchButtons: this.showSwitchButtons,
-          }
-    },
-    currentProperties(): GeoJsonProperties {
-      const properties = {
-        ...this.windowFeatures[this.visibleWindowFeatureIndex],
-      }
-      const exportProperty =
-        this.exportPropertyLayerKeys[properties.polarInternalLayerKey]
-      if (exportProperty?.length > 0) {
-        delete properties[exportProperty]
-      }
-      return properties
-    },
-    exportProperty(): string {
-      if (this.currentProperties) {
-        const property =
-          this.exportPropertyLayerKeys[
-            this.currentProperties.polarInternalLayerKey
-          ]
-        return property?.length > 0
-          ? this.windowFeatures[this.visibleWindowFeatureIndex][property]
-          : ''
-      }
-      return ''
+      return this.showList ? {} : {}
     },
     moveHandleProperties() {
       const properties: MoveHandleProperties = {
@@ -89,10 +58,6 @@ export default Vue.extend({
     },
     renderUi(): boolean {
       return this.windowFeatures.length > 0 || this.showList
-    },
-    /** only show switch buttons if multiple property sets are available */
-    showSwitchButtons(): boolean {
-      return this.windowFeatures.length > 1
     },
   },
   watch: {
@@ -114,13 +79,6 @@ export default Vue.extend({
       }
     },
   },
-  mounted() {
-    window.addEventListener('resize', this.updateClientWidth)
-    this.updateClientWidth()
-  },
-  beforeDestroy() {
-    window.removeEventListener('resize', this.updateClientWidth)
-  },
   methods: {
     ...mapMutations(['setMoveHandle']),
     ...mapActions('plugin/gfi', ['close']),
@@ -140,9 +98,6 @@ export default Vue.extend({
           )
         )
       }
-    },
-    updateClientWidth() {
-      this.clientWidth = this.$root.$el.clientWidth
     },
   },
 })